reload page after authenticating
cas assigns a ticket parameter when redirecting to the application which is validated by the application. If someone reloads the page (not knowing about the ticket parameter), the application will validate the ticket and fail since the ticket was already validated. Reloading the page after saving the user to the current session removes the ticket query parameter.

We could also check if the user is already in the current session before validating the ticket, but it's cleaner just to remove the ticket parameter completely.